return early from grade_user() if grader_endpoint is invalid
This moves the check on whether `grader_endpoint` is a valid URL to
the beginning of `grade_user()`. This achieves two things:

1. if the endpoint is not a valid URL, there's no point in attempting
to do anything else, like hit the authentication endpoint. Just fail
with an error message immediately. That also lets us remove an
indentation level from most of the method, which helps readability.
2. the check was only being applied to `get` requests. The
`http_method == "post"` case is now also covered by this, so it will
also return a proper error message instead of just attempting to
request an invalid URL.